Only update files if they've changed

+++ artsy.py
@@ -4,8 +4,9 @@
 import os
 import shutil
 import json
+import binascii
+import glob
 from collections import OrderedDict
-from distutils.dir_util import copy_tree
 from resizeimage import resizeimage
 from PIL import Image
 import cattr
@@ -16,12 +17,28 @@ from templater import Templater
 templater = Templater("templates")
 
 
+def get_hash(infile):
+    if not os.path.exists(infile):
+        return None
+
+    return "%08X" % (binascii.crc32(open(infile, "rb").read()) & 0xFFFFFFFF)
+
+
+def remove_parent_path(parent, filepath):
+    return filepath[len(parent) + 1:]
+
+
+def get_dir_hashes(indir):
+    files = glob.iglob("{}/**".format(indir), recursive=True)
+    return {remove_parent_path(indir, f): get_hash(f) for f in files if os.path.isfile(f)}
+
+
 def generate_thumbnail_size(img, width, filename):
     thumb = resizeimage.resize_thumbnail(img, [width, width])
     thumb.save(filename, img.format)
 
 
-def generate_thumbnails(indir, outdir, image, widths):
+def generate_thumbnails(indir, outdir, image, widths, do_update, add_touched, force=False):
     thumbnails = {}
     relpath = os.path.join(indir, image.filename)
 
@@ -30,15 +47,23 @@ def generate_thumbnails(indir, outdir, image, widths):
 
     # Copy full image file
     fullpath = get_path("full")
-    shutil.copy2(relpath, fullpath)
+    fullhash = get_hash(relpath)
+    should_do_update = do_update(fullpath, fullhash)
+    if should_do_update:
+        shutil.copy2(relpath, fullpath)
+    add_touched(fullpath)
+
     thumbnails["full"] = os.path.basename(fullpath)
+    thumbnails["_relpath"] = relpath
 
     # Generate actual thumbnails
     with Image.open(relpath) as img:
         for width in list(widths):
             filename = get_path(width)
-            generate_thumbnail_size(img, width, filename)
+            if should_do_update or not os.path.exists(filename):
+                generate_thumbnail_size(img, width, filename)
             thumbnails[width] = os.path.basename(filename)
+            add_touched(filename)
 
     return thumbnails
 
@@ -58,24 +83,58 @@ def clean_empty(d):
     return {k: v for k, v in ((k, clean_empty(v)) for k, v in d.items()) if v}
 
 
-def write_json(outfile, data):
-    destdata = clean_empty(cattr.unstructure(data))
+def write_json(outfile, data, clean=False):
+    if clean:
+        data = clean_empty(cattr.unstructure(data))
+
     with open(outfile, "w") as f:
-        json.dump(destdata, f)
+        json.dump(data, f)
+
+
+def cleanup_dead_files(output_dir, tree_hash, touched_files):
+    dead_files = set(tree_hash.keys()) - set(touched_files)
+
+    for df in dead_files:
+        print("Deleting file {}".format(df))
+        os.remove(os.path.join(output_dir, df))
 
 
 # TODO: Don't update files that don't need updating
-def generate_static_site(input_dir, output_dir, limit="*"):
+def generate_static_site(input_dir, output_dir, limit="*", force=False):
     '''Output templates to filesystem.'''
     # Get data and fail on error
     data = get_art_data(input_dir, limit)
 
-    # Recreate the output directory
-    if os.path.exists(output_dir):
-        shutil.rmtree(output_dir, ignore_errors=True)
+    touched_files = []
 
-    # Copy static files (and create output dir)
-    copy_tree("static", output_dir)
+    def add_touched(filename):
+        touched_files.append(remove_parent_path(output_dir, filename))
+
+    def do_update(fullpath, fullhash):
+        return force or not os.path.exists(fullpath) or fullpath not in tree_hash or fullhash != tree_hash[fullpath]
+
+    tree_hash = {}
+
+    if os.path.exists(output_dir):
+        if force:
+            # Recreate the output directory
+            shutil.rmtree(output_dir, ignore_errors=True)
+
+        # Get all current hashes
+        tree_hash = get_dir_hashes(output_dir)
+    else:
+        # Create output dir
+        os.makedirs(output_dir)
+
+    # Copy static files
+    static_files = glob.iglob("static/**", recursive=True)
+    for filepath in [item for item in static_files if os.path.isfile(item)]:
+        filehash = get_hash(filepath)
+        newpath = remove_parent_path("static", filepath)
+        outfile = os.path.join(output_dir, newpath)
+        if do_update(outfile, filehash):
+            shutil.copy2(filepath, outfile)
+        add_touched(outfile)
 
     # Hold thumbnail paths
     thumbnails = {}
@@ -105,10 +164,11 @@ def generate_static_site(input_dir, output_dir, limit="*"):
             outfile = os.path.join(artistdir, "{}.html".format(image.slug))
 
             # Generate thumbnails
-            thumbnails[image.slug] = generate_thumbnails(artist_file._reldir, artistdir, image, [120, 512])
+            thumbnails[image.slug] = generate_thumbnails(artist_file._reldir, artistdir, image, [120, 512], do_update, add_touched)
 
             # Write templated file
             write_page("image", outfile, artist=artist, image=image, get_character_breakdown=data.get_character_breakdown, get_species_details=data.get_species_details, get_tag_details=data.get_tag_details, thumbnails=thumbnails[image.slug])
+            add_touched(outfile)
 
             artistout["files"].append(image)
 
@@ -117,12 +177,14 @@ def generate_static_site(input_dir, output_dir, limit="*"):
 
         # Write templated file
         write_page("artist", artistfile, artist=artist, files=files, thumbnails=thumbnails)
+        add_touched(artistfile)
 
         use_artists.append(artistout)
 
     # Generate all-artists template
     artistsfile = os.path.join(output_dir, "all_artists.html")
     write_page("artists", artistsfile, artists=use_artists, limit=limit, thumbnails=thumbnails)
+    add_touched(artistsfile)
 
     # Generate tag templates
     use_tags = list(sorted(data.get_all_tags(limit=limit, ignore="species")))
@@ -138,10 +200,12 @@ def generate_static_site(input_dir, output_dir, limit="*"):
 
         outfile = os.path.join(tagsdir, "{}.html".format(t.replace("#", "_")))
         write_page("tag", outfile, tag=t, description=data.get_tag_details(t), files=files, thumbnails=thumbnails)
+        add_touched(outfile)
 
     # Generate all-tags template
     tagfile = os.path.join(output_dir, "all_tags.html")
     write_page("tags", tagfile, tags=collected_tags, thumbnails=thumbnails, get_tag_details=data.get_tag_details)
+    add_touched(tagfile)
 
     # Generate species templates
     use_species = list(sorted(data.get_all_species(limit=limit)))
@@ -149,7 +213,7 @@ def generate_static_site(input_dir, output_dir, limit="*"):
     specdir = os.path.join(output_dir, "_species")
     if not os.path.exists(specdir):
         os.makedirs(specdir)
-    
+
     collected_species = OrderedDict()
     for spec in use_species:
         files = data.get_files_by_species(spec, limit=limit)
@@ -157,10 +221,12 @@ def generate_static_site(input_dir, output_dir, limit="*"):
 
         outfile = os.path.join(specdir, "{}.html".format(spec))
         write_page("species", outfile, species=spec, description=data.get_species_details(spec), files=files, thumbnails=thumbnails)
-    
+        add_touched(outfile)
+
     # Generate all-species template
     specfile = os.path.join(output_dir, "all_species.html")
     write_page("species_all", specfile, species=collected_species, thumbnails=thumbnails, get_species_details=data.get_species_details)
+    add_touched(specfile)
 
     # Collect unique characters
     # TODO: Sort this by the character metadata definition
@@ -184,20 +250,22 @@ def generate_static_site(input_dir, output_dir, limit="*"):
 
             if len(sf) > 0 or s.subforms:
                 thischarspec[sn] = {"species": s, "files": sf, "subforms": {}}
-            
+
             if s.subforms:
                 for sfn, sf in s.subforms.items():
                     sff = data.get_files_by_character(char_name, species=sn, subform=sfn, limit=limit)
 
                     if len(sff) > 0:
                         thischarspec[sn]["subforms"][sfn] = {"subform": sf, "files": sff}
-            
+
         outfile = os.path.join(chardir, "{}.html".format(char_name))
         write_page("character", outfile, data=thischar, thumbnails=thumbnails, get_species_details=data.get_species_details)
+        add_touched(outfile)
 
     # Generate all-characters template
     charfile = os.path.join(output_dir, "all_characters.html")
     write_page("characters", charfile, characters=collected_chars, thumbnails=thumbnails)
+    add_touched(charfile)
 
     # Generate JSON file
     jsonfile = os.path.join(output_dir, "data.json")
@@ -211,7 +279,8 @@ def generate_static_site(input_dir, output_dir, limit="*"):
         "thumbnails": thumbnails
     }
 
-    write_json(jsonfile, jsondata)
+    write_json(jsonfile, jsondata, True)
+    add_touched(jsonfile)
 
     # Generate index file
     indexfile = os.path.join(output_dir, "index.html")
@@ -220,17 +289,20 @@ def generate_static_site(input_dir, output_dir, limit="*"):
         "thumbnails": thumbnails,
         "files": data.get_all_files(limit=limit),  # Called manually here to sort the entire thing
         "artists": map(lambda x: x["artist"], use_artists),
-        "tags": jsondata["tags"],
-        "species": jsondata["species"],
+        "tags": use_tags,
+        "species": use_species,
         "characters": list(sorted(chars)),
-        "character_data": jsondata["characters"],
+        "character_data": use_chars,
         "get_tag_details": data.get_tag_details,
         "get_species_details": data.get_species_details
     }
 
     write_page("index", indexfile, **indexdata)
+    add_touched(indexfile)
+
+    cleanup_dead_files(output_dir, tree_hash, touched_files)
 
 
 if __name__ == "__main__":
-    generate_static_site("testdata", "output")
+    generate_static_site("testdata", "output", force=True)
     print("Files written.")
